Optimization

Optimization techniques

SRBF

Surrogates.surrogate_optimize — Method.

The main idea is to pick the new evaluations from a set of candidate points where each candidate point is generated as an N(0, sigma^2) distributed perturbation from the current best solution. The value of sigma is modified based on progress and follows the same logic as in many trust region methods; we increase sigma if we make a lot of progress (the surrogate is accurate) and decrease sigma when we aren’t able to make progress (the surrogate model is inaccurate). More details about how sigma is updated is given in the original papers.

After generating the candidate points we predict their objective function value and compute the minimum distance to previously evaluated point. Let the candidate points be denoted by C and let the function value predictions be s(x_i) and the distance values be d(x_i), both rescaled through a linear transformation to the interval [0,1]. This is done to put the values on the same scale. The next point selected for evaluation is the candidate point x that minimizes the weighted-distance merit function:

$merit(x) = ws(x) + (1-w)(1-d(x))$

where $0 \leq w \leq 1$. That is, we want a small function value prediction and a large minimum distance from previously evalauted points. The weight w is commonly cycled between a few values to achieve both exploitation and exploration. When w is close to zero we do pure exploration while w close to 1 corresponds to explotation.

source

LCBS

Surrogates.surrogate_optimize — Method.

This is an implementation of Lower Confidence Bound (LCB), a popular acquisition function in Bayesian optimization. Under a Gaussian process (GP) prior, the goal is to minimize:

$LCB(x) := E[x] - k * \sqrt{(V[x])}$

default value $k = 2$.

source

EI

Surrogates.surrogate_optimize — Method.

This is an implementation of Expected Improvement (EI), arguably the most popular acquisition function in Bayesian optimization. Under a Gaussian process (GP) prior, the goal is to maximize expected improvement:

$EI(x) := E[max(f_{best}-f(x),0)$

source

DYCORS

Missing docstring.

Missing docstring for surrogate_optimize(obj::Function,::DYCORS,lb,ub,krig,sample_type::SamplingAlgorithm;maxiters=100,num_new_samples=100). Check Documenter's build log for details.

Adding another optimization method

To add another optimization method, you just need to define a new SurrogateOptimizationAlgorithm and write its corresponding algorithm, overloading the following:

surrogate_optimize(obj::Function,::NewOptimizatonType,lb,ub,surr::AbstractSurrogate,sample_type::SamplingAlgorithm;maxiters=100,num_new_samples=100)