Samples

The syntax for sampling in an interval or region is the following:

sample(n,lb,ub,S::SamplingAlgorithm)

where lb and ub are, respectively, the lower and upper bounds. There are many sampling algorithms to choose from:

Grid sample

Surrogates.GridSample — Type

GridSample{T}

T is the step dx for lb:dx:ub

Surrogates.sample — Method

sample(n,lb,ub,S::GridSample)

Returns a tuple containing numbers in a grid.

Uniform sample

Surrogates.sample — Method

sample(n,lb,ub,::UniformRandom)

Returns a Tuple containing uniform random numbers.

Sobol sample

Surrogates.sample — Method

sample(n,lb,ub,::SobolSampling)

Returns a Tuple containing Sobol sequences.

Latin Hypercube sample

Surrogates.sample — Method

sample(n,lb,ub,::LatinHypercube)

Returns a Tuple containing LatinHypercube sequences.

Low Discrepancy sample

Surrogates.LowDiscrepancySample — Type

LowDiscrepancySample{T}

T is the base for the sequence

Surrogates.sample — Method

sample(n,lb,ub,S::LowDiscrepancySample)

Low discrepancy sample:

Dimension 1: Van der Corput sequence

Dimension > 1: Halton sequence

If dimension d > 1, all bases must be coprime with each other.

Adding a new sampling method

Adding a new sampling method is a two- step process:

Adding a new SamplingAlgorithm type

Overloading the sample function with the new type.

Example

struct NewAmazingSamplingAlgorithm{OPTIONAL} <: SamplingAlgorithm end

function sample(n,lb,ub,::NewAmazingSamplingAlgorithm)
    if lb is  Number
        ...
        return x
    else
        ...
        return Tuple.(x)
    end
end